+++ nft-studio/src/components/InvestRedeem.tsx
@@ -22,7 +22,7 @@ import styled from 'styled-components'
 import { Dec } from '../utils/Decimal'
 import { formatBalance, getCurrencySymbol } from '../utils/formatting'
 import { useAddress } from '../utils/useAddress'
-import { useBalances } from '../utils/useBalances'
+import { getBalanceDec, useBalances } from '../utils/useBalances'
 import { useCentrifugeTransaction } from '../utils/useCentrifugeTransaction'
 import { usePermissions } from '../utils/usePermissions'
 import { usePendingCollect, usePool, usePoolMetadata } from '../utils/usePools'
@@ -45,17 +45,6 @@ export const InvestRedeem: React.VFC<Props> = (props) => {
       <InvestRedeemInner {...props} />
     </LoadBoundary>
   )
-}
-
-type Balances = Exclude<ReturnType<typeof useBalances>, undefined>
-
-function getBalanceDec(balances: Balances, currency: string) {
-  if (currency === 'native') {
-    return Dec(balances.native.balance.toString()).div(Dec(10).pow(balances.native.decimals))
-  }
-  const entry = balances.currencies.find((c) => c.currency === currency)
-  if (!entry) throw new Error(`invalid currency: ${currency}`)
-  return entry.balance.toDecimal()
 }
 
 // function min(...nums: Decimal[]) {
