+++ nft-studio/src/utils/useBalances.ts
@@ -1,3 +1,4 @@
+import { Dec } from './Decimal'
 import { useCentrifugeQuery } from './useCentrifugeQuery'
 
 export function useBalances(address?: string) {
@@ -6,4 +7,15 @@ export function useBalances(address?: string) {
   })
 
   return result
+}
+
+type Balances = Exclude<ReturnType<typeof useBalances>, undefined>
+
+export function getBalanceDec(balances: Balances, currency: string) {
+  if (currency === 'native') {
+    return Dec(balances.native.balance.toString()).div(Dec(10).pow(balances.native.decimals))
+  }
+  const entry = balances.currencies.find((c) => c.currency === currency)
+  if (!entry) throw new Error(`invalid currency: ${currency}`)
+  return entry.balance.toDecimal()
 }
